refactor: 코틀린 서버 폴더 수정 및 gradle 설정 추가
